qwerty_webapp/requirements.txt
flet>=0.23
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.7
